idl: support `const` specifier member template parameters
This patch introduces support for the following syntax in IDL files:

    struct my_struct {
        std::vector<const my_struct_2> member;
    };

No significant code generation changes involved except for patching
static_assert:s to propagate correctly `const` in template parameters.

The following syntax is not supported, though:

    struct my_struct {
        const std::vector<my_struct_2> member;
    };

Also parser now supports `const` class members with non-template type,
e.g.:

    struct my_struct {
        const my_struct_2 member;
    };

But in this case the use is limited only to the classes without
`[[writable]]` attribute.

The reason for the limited support is that the current code is
quite hard to maintain in the current state, parsing needs to be
reworked towards generating an AST and need to manipulate on
synthesized AST nodes instead of raw token lists. This is out of
scope for the current change, though.

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>