raft: make `log.last_conf_idx()` to always point at a valid configuration
Prior to the fix `log.last_conf_idx()` returned 0 when
the latest valid  configuration should be
obtained from a snapshot (e.g. when there are no
configuration entries in the log, yet).

`log.get_configuration()` already accounted for it by
returning configuration from a snapshot in case
`_last_conf_idx == 0`. Do the same for `log.last_conf_idx()`
so that `last_conf_idx()` always points to a valid
configuration entry, either from the log (if it's available)
or a snapshot.

Tests: unit(dev)

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>