cql_query_test: test behavior of `uuid()` and `timeuuid*()` functions wrt prepared statements
`uuid()` CQL function isn't marked as "pure" hence
its execution is delayed until the actual execution
of the function begins, which seems to be logically
correct.
Some other functions, related to timeuuid, e.g.
`currenttimeuuid()`, are marked as "pure" so that
they are executed when a statement is being prepared.

This can lead to unobvious behavior if used in
partition keys with insert statement, for example.
The former case would create a new partition each
time a prepared statement is executed, but the latter
will not and could possibly cause overwriting data
in the same partition.

Add a test case to reproduce the issue.

Tests: unit(dev)

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>