db: add experimental option for raft schema changes
Introduce `raft-schema-changes` experimental option.
Adjust the tests accordingly to accomodate the new option.

Don't wire it up now, because the code for raft-based
schema changes is still in an early stage of development.
Also, don't introduce a corresponding gossiper feature
just yet, because again, it should be done after the
raft schema changes API contract is stabilized.

This will be done in a separate series, probably related to
implementing the feature itself.

Tests: unit(dev)

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>